reorganizing
1. removed freemium files in the API folder and added them to helper.ts.  did the same with the tests.  I had to keep the .tsx file because it has components and cannot be included in a ts file.
I also hard coded the TOTAL_TRIAL DAYS in here and imported it to necessary locations.
